milgiorato definitivamente profile.py con commenti

+++ app/blueprints/profile.py
@@ -1,65 +1,66 @@
 from sqlite3 import IntegrityError
-from flask import Blueprint, render_template, redirect, url_for, request, session, flash
+from flask import Blueprint, abort, render_template, redirect, url_for, request, session, flash
 from flask_login import login_required, fresh_login_required, current_user, logout_user
-from extensions.database import Address, Cart, Notification, Order, OrderProduct, Profile, SellerInformation, User, Product, Role, Category, db
+from extensions.database import Address, Cart, Notification, Order, OrderProduct, Profile, SellerInformation, State, User, Product, Role, Category, db
 from extensions.princ import buyer_required, admin_required, admin_permission, buyer_permission
 from werkzeug.security import check_password_hash, generate_password_hash
 from datetime import date
 
 app = Blueprint('profile', __name__)
 
+# Gestisce la selezione del profilo dopo il login
 @app.route('/profile/select', methods = ['GET'])
 @login_required
 @fresh_login_required
 def select():
     return render_template('profile_selection.html')
 
+# Gestisce l'indirizzamento dopo la scelta del profilo
 @app.route('/profile/select/<int:profile_id>/<int:action>', methods = ['GET'])
 @login_required
 @fresh_login_required
 def select_id(profile_id, action):
+    # Action deve corrispondere a determinati valori
+    if action != 0 and  action != 1:
+        abort(404)
     try:
+        # Cerco se il profilo corrente corrisponde a un profilo associato all'utente
         profile = next((p for p in current_user.profiles if p.id == profile_id),None)
-        if not profile:                
-            flash('Il profilo selezionato non esiste', "error")
-            return redirect(url_for('shop.shop'))
+        if not profile or not profile.is_valid:                
+            flash('Profilo non trovato o non caricato correttamente', 'error')
+            return redirect(url_for('auth.logout'))
 
-        # Here you can set the selected profile in the session or any other logic
+        # Setto il profilo corrente che sta operando
         session['current_profile_id'] = profile.id
-        flash('Profilo aggiornato con successo', "success")
-        if action == 0:
-            return redirect(url_for('shop.shop'))
-        elif action == 1:
-            return redirect(url_for('shop.shop'))
-        elif action == 2:
-            return redirect(url_for('cart.cart'))
-        else:
-            return redirect(url_for('auth.logout'))
     except Exception:
-        flash('Si è verificato un errore', "error")
-        if action == 0:
-            return redirect(url_for('shop.shop'))
-        elif action == 1:
-            return redirect(url_for('shop.shop'))
-        elif action == 2:
-            return redirect(url_for('cart.cart'))
-        else:
-            return redirect(url_for('auth.logout'))
+        flash("Si è verificato un errore di database. Riprova più tardi", "error")
+        return redirect(url_for('auth.logout'))
+    
+    flash('Profilo aggiornato correttamente', "success")
+    if action == 0:
+        return redirect(url_for('shop.shop'))
+    else:
+        return redirect(url_for('shop.shop'))
 
+# Gestisce la creazione di un nuovo profilo
 @app.route('/profile/add/<int:action>', methods=['GET', 'POST'])
 @login_required
 @fresh_login_required
-def add(action):  
+def add(action):
+    # Action deve corrispondere a determinati valori
+    if action != 0 and action != 1:
+        abort(404)
+
     if request.method == 'POST':
         name = request.form.get('name')
         surname = request.form.get('surname')
-        birth_date = (request.form.get('birth_date'))
-
+        birth_date = request.form.get('birth_date')
+        
+        # Controllo i dati del form
         if not name or not surname or not birth_date:
-            flash('Inserisci tutti i campi', 'fail')
-            return render_template('add_profile.html', action = action)
-        # Begin database session
-        #db.session.begin()
+            flash('Inserisci tutti i campi', 'error')
+            return redirect(url_for('profile.add', action = action))
+        
         try:
             # Ottieni la data corrente
             current_date = date.today()
@@ -77,55 +78,60 @@ def add(action):
             if birth_date < earliest_date:
                 flash(f"L'anno di nascita deve essere compreso tra {earliest_date.year} e {current_date.year}.")
                 return redirect(url_for('profile.add', action = action))
-            # Create new profile
+            
             new_profile = Profile(name=name, surname=surname, birth_date=birth_date, user_id=current_user.id)
             db.session.add(new_profile)
             db.session.commit()
-            flash('Profilo aggiunto correttamente', 'success')
-            if action == 0:
-                return redirect(url_for('profile.select'))
-            elif action == 1:
-                return redirect(url_for('account.view'))
+
         except IntegrityError:
-            # Handle unique constraint violation (e.g., duplicate profile name)
             db.session.rollback()
-            flash('Un porfilo con lo stesso nome esiste già', "error")
+            flash('Profilo già esistente', "error")
             if action == 0:
                 return redirect(url_for('profile.select'))
-            elif action == 1:
+            else:
                 return redirect(url_for('account.view'))
-        except Exception as e:
-            print(e)
-            # Catch other unexpected errors
+        except Exception:
             db.session.rollback()
             flash('Si è verificato un errore di database. Riprova più tardi.', "error")
             if action == 0:
                 return redirect(url_for('profile.select'))
-            elif action == 1:
+            else:
                 return redirect(url_for('account.view'))
+            
+        flash('Profilo aggiunto correttamente', 'success')
+        if action == 0:
+            return redirect(url_for('profile.select'))
+        else:
+            return redirect(url_for('account.view'))
         
     return render_template('add_profile.html', action=action)
 
+# Gestisce la modifica di un profilo dell'utente
 @app.route('/profile/modify/<int:profile_id>', methods=['GET', 'POST'])
 @login_required
 @fresh_login_required
 def modify(profile_id):
 
-    profile = next((p for p in current_user.profiles if p.id == profile_id), None)
-
-    if not profile:
-        flash('Il profilo non è stato trovato', "error")
-        return redirect(url_for('account.view'))
+    try:
+        # Cerco se il profilo corrente corrisponde a un profilo associato all'utente
+        profile = next((p for p in current_user.profiles if p.id == profile_id),None)
+        if not profile or not profile.is_valid:                
+            flash('Profilo non trovato o non caricato correttamente', 'error')
+            return redirect(url_for('auth.logout'))
         
+    except Exception:
+        flash('Si è verificato un errore di database. Riprova più tardi','error')
+        return redirect(url_for('auth.logout'))
+    
     if request.method == 'POST':
-        # Leggi i dati inviati dal form
         name = request.form.get('name')
         surname = request.form.get('surname')
         image_url = request.form.get('image_url')
-        birth_date = (request.form.get('birth_date'))
+        birth_date = request.form.get('birth_date')
 
+        # Controllo dei dati del form
         if not name or not surname or not image_url or not birth_date:
-            flash('Inserisci tutti i dati richiesti', 'warning')
+            flash('Inserisci tutti i dati richiesti', 'error')
             return redirect(url_for('modify_profile', profile_id=profile_id))
         
         try:
@@ -152,72 +158,70 @@ def modify(profile_id):
             profile.image_url = image_url
 
             db.session.commit()
-            
-            flash('Profilo aggiornato con successo')
-            return redirect(url_for('account.view'))
 
         except ValueError:
             flash('Formato della data non valido.',"error")
             return redirect(url_for('profile.modify', profile_id=profile_id))
-        except Exception as e:
-            print(f"Errore durante l'operazione: {str(e)}")
-            print('ciao')
+        except Exception:
             db.session.rollback()
             flash('Si è verificato un errore di database. Riprova più tardi.', "error")
             return redirect(url_for('account.view'))
+        
+        flash('Profilo aggiornato con successo')
+        return redirect(url_for('account.view'))
     
     return render_template('modify_profile.html', profile=profile)
 
+# Gestisce l'eliminazione del profilo e di tutte le sue informazioni
 @app.route('/profile/delete/<int:profile_id>', methods=['GET'])
 @login_required
 @fresh_login_required
 def delete(profile_id):
-    # Begin database session
-    #db.session.begin()
     try:
-        profile = next((p for p in current_user.profiles if p.id == profile_id), None)
-        
-        if not profile:
-            flash('Il profilo non è stato trovato', "error")
-            return redirect(url_for('account.view'))
+        # Cerco se il profilo corrente corrisponde a un profilo associato all'utente
+        profile = next((p for p in current_user.profiles if p.id == profile_id),None)
+        if not profile or not profile.is_valid:                
+            flash('Profilo non trovato o non caricato correttamente', 'error')
+            return redirect(url_for('auth.logout'))
         
         # Se è presente un solo profilo, allora non posso eliminarlo, altrimenti non avrei un profilo con cui navigare lo shop
-        if len(current_user.profiles) > 1:
-            for item in profile.cart_items:
-                db.session.delete(item)
-                
-            db.session.delete(profile)
-
-            db.session.commit()
-            if profile.id == session['current_profile_id']:
-                session['current_profile_id'] = current_user.profiles[0].id
-            flash('Profilo eliminato correttamente', "success")
-            return redirect(url_for('account.view')) 
-        else:
-            flash("Non puoi eliminare l'unico profilo rimanente.", 'fail')
+        if len(current_user.profiles) <= 1:
+            flash("Non puoi eliminare l'unico profilo rimanente.", 'error')
             return redirect(url_for('account.view'))
-    except Exception as e:
+        
+        # Elimino tutti gli item, associati a quel prodotto, dei carrelli degli altri utenti 
+        for item in profile.cart_items:
+            db.session.delete(item)
+
+        # Elimino il prodilo
+        db.session.delete(profile)
+
+        # Se il profilo eliminato corrispondeva al profilo corrente, allora associa il profilo corrente al primo profilo disponibile
+        if profile.id == session['current_profile_id']:
+            session['current_profile_id'] = current_user.profiles[0].id
+        db.session.commit()
+
+    except Exception:
         db.session.rollback()
-        print(str(e))
         flash('Si è verificato un errore di database. Riprova più tardi.', "error")
         return redirect(url_for('account.view'))
+    
+    flash('Profilo eliminato correttamente', "success")
+    return redirect(url_for('account.view')) 
 
+# Gestisce la visualizzazione di tutte le informazioni del database(SOLO PER L'ADMIN)
 @app.route('/profile/info', methods = ['GET'])
 @login_required
 @fresh_login_required
-#@admin_required
-#@permission_required(buyer_permission)
+@admin_required
 def info():
-    # Ricaviamo tutti gli utenti della tabella User, tutti i prodotti di Products e tutti i Ruoli
+    # Ricaviamo tutti gli utenti della tabella User, e tutte le tabelle necessarie affinchè il database funzioni correttamente
     try:
-        all_products = Product.query.all()
-        all_orders = Order.query.all()
-        all_notifications = Notification.query.all()
         all_users = User.query.all()
         all_roles = Role.query.all()
         all_categories = Category.query.all()
-        seller_information = SellerInformation.query.all()
+        all_states = State.query.all()
     except Exception:
-        #flash('Si è verificato un errore di database. Riprova più tardi.', "error")
-        flash('La pagina info.html non è stata caricata correttamente',"error")
-    return render_template('info.html', products = all_products, orders = all_orders, notifications = all_notifications, users=all_users, roles=all_roles, categories=all_categories, seller_information = seller_information) # Passo anche lo username dell'utente loggato(sarà sempre unico)
+        flash('Si è verificato un errore di database. Riprova più tardi.', "error")
+        return redirect(url_for('account.view'))
+    return render_template('info.html', users=all_users, roles=all_roles, categories=all_categories, states = all_states)
